Refactor - remove duplicate TargetPlayerCardsInHand classes and made CardsInTargetHandCount shared
