+++ Mage.Sets/src/mage/sets/seventhedition/SuddenImpact.java
@@ -27,17 +27,15 @@
  */
 package mage.sets.seventhedition;
 
-import java.util.UUID;
-import mage.abilities.Ability;
-import mage.abilities.dynamicvalue.DynamicValue;
+import mage.abilities.dynamicvalue.common.CardsInTargetHandCount;
 import mage.abilities.effects.Effect;
 import mage.abilities.effects.common.DamageTargetEffect;
 import mage.cards.CardImpl;
 import mage.constants.CardType;
 import mage.constants.Rarity;
-import mage.game.Game;
-import mage.players.Player;
 import mage.target.TargetPlayer;
+
+import java.util.UUID;
 
 /**
  *
@@ -51,7 +49,7 @@ public class SuddenImpact extends CardImpl {
 
 
         // Sudden Impact deals damage to target player equal to the number of cards in that player's hand.
-        Effect effect = new DamageTargetEffect(new TargetPlayerCardsInHandCount());
+        Effect effect = new DamageTargetEffect(new CardsInTargetHandCount());
         effect.setText("{this} deals damage to target player equal to the number of cards in that player's hand.");
         this.getSpellAbility().addEffect(effect);
         this.getSpellAbility().addTarget(new TargetPlayer());
@@ -64,33 +62,5 @@ public class SuddenImpact extends CardImpl {
     @Override
     public SuddenImpact copy() {
         return new SuddenImpact(this);
-    }
-}
-
-class TargetPlayerCardsInHandCount implements DynamicValue {
-
-    @Override
-    public int calculate(Game game, Ability sourceAbility, Effect effect) {
-        Player targetPlayer = game.getPlayer(sourceAbility.getFirstTarget());
-        if (targetPlayer != null) {
-            return targetPlayer.getHand().size();
-        }
-
-        return 0;
-    }
-
-    @Override
-    public DynamicValue copy() {
-        return new TargetPlayerCardsInHandCount();
-    }
-
-    @Override
-    public String toString() {
-        return "1";
-    }
-
-    @Override
-    public String getMessage() {
-        return "target player's cards in hand";
     }
 }
